+++ src/components/shapes/fractions/index.js
@@ -1,5 +1,6 @@
 // src/components/shapes/fractions/index.js
 export { default as DiskFraction } from "./DiskFraction";
+
 export { default as SquareFraction } from "./SquareFraction";
 export { default as SquareDiagonalFraction } from "./SquareDiagonalFraction";
 export { default as SquareCornerTriangleFraction } from "./SquareCornerTriangleFraction";
@@ -8,9 +9,12 @@ export { default as SquareCrossFraction } from "./SquareCrossFraction";
 export { default as SquareIsoscelesTriangleFraction } from "./SquareIsoscelesTriangleFraction";
 export { default as SquareRectangleThin8thFraction } from "./SquareRectangleThin8thFraction";
 export { default as SquareHalfRectangle8thFraction } from "./SquareHalfRectangle8thFraction";
+
 export { default as RectangleFraction } from "./RectangleFraction";
 export { default as RectangleDiagonalFraction } from "./RectangleDiagonalFraction";
 export { default as RectangleQuarterMedianFraction } from "./RectangleQuarterMedianFraction";
 export { default as RectangleQuarterQuadrantFraction } from "./RectangleQuarterQuadrantFraction";
 export { default as RectangleQuarterRectangleFraction } from "./RectangleQuarterRectangleFraction";
+export { default as RectangleLShapeFraction } from "./RectangleLShapeFraction";
+
 export { default as HouseFraction } from "./HouseFraction";